Similarity_python/requirements.txt
# Web Framework
flask==3.1.1
flask-cors==6.0.0

# Database
pymongo==4.10.1

# Video / Audio Processing
opencv-python==4.11.0.86
numpy==1.26.4
scipy==1.16.3

# NLP / Embeddings
spacy==3.8.11
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
sentence-transformers==3.4.1
scikit-learn==1.7.2

# Speech-to-Text
faster-whisper==1.1.1

# Computer Vision / Face Analysis
mediapipe==0.10.21
deepface==0.0.93

# ML Framework
tensorflow-cpu==2.19.0
tf-keras==2.19.0
torch==2.7.1
torchvision==0.22.1

# HTTP / API
requests==2.32.3

# Utilities
python-dotenv==1.1.1
